Integrate downstream changes, from PR #424
