+++ nucypher/utilities/sandbox.py
@@ -60,6 +60,7 @@ def make_ursulas(ether_addresses: list,
     if isinstance(ether_addresses, int):
         ether_addresses = [to_checksum_address(secure_random(20)) for _ in range(ether_addresses)]
 
+
     event_loop = asyncio.get_event_loop()
     if not _TEST_KNOWN_URSULAS_CACHE:
         starting_port = constants.URSULA_PORT_SEED
@@ -97,8 +98,6 @@ def make_ursulas(ether_addresses: list,
                             federated_only=federated_only,
                             **ursula_kwargs)
 
-            ursula.attach_rest_server()
-
             class MockDatastoreThreadPool(object):
                 def callInThread(self, f, *args, **kwargs):
                     return f(*args, **kwargs)
@@ -121,7 +120,10 @@ def make_ursulas(ether_addresses: list,
             ursula.federated_only = True
 
         ursulas.add(ursula)
-        _TEST_KNOWN_URSULAS_CACHE[ursula.rest_interface.port] = ursula
+
+        # Store this Ursula in our global cache.
+        port = ursula.rest_information()[0].port
+        _TEST_KNOWN_URSULAS_CACHE[port] = ursula
 
     if know_each_other and not bare:
 
@@ -192,6 +194,14 @@ class MockRestMiddleware(RestMiddleware):
     class NotEnoughMockUrsulas(MinerAgent.NotEnoughMiners):
         pass
 
+    def __get_mock_client_by_ursula(self, ursula):
+        port = ursula.rest_information()[0].port
+        return self.__get_mock_client_by_port(port)
+
+    def __get_mock_client_by_url(self, url):
+        port = int(url.split(":")[1])
+        return self.__get_mock_client_by_port(port)
+
     def __get_mock_client_by_port(self, port):  # TODO
         try:
             ursula = _TEST_KNOWN_URSULAS_CACHE[port]
@@ -203,25 +213,25 @@ class MockRestMiddleware(RestMiddleware):
         return mock_client
 
     def consider_arrangement(self, arrangement=None):
-        mock_client = self.__get_mock_client_by_port(arrangement.ursula.rest_interface.port)
+        mock_client = self.__get_mock_client_by_ursula(arrangement.ursula)
         response = mock_client.post("http://localhost/consider_arrangement", bytes(arrangement))
         assert response.status_code == 200
         return response
 
     def enact_policy(self, ursula, id, payload):
-        mock_client = self.__get_mock_client_by_port(ursula.rest_interface.port)
+        mock_client = self.__get_mock_client_by_ursula(ursula)
         response = mock_client.post('http://localhost/kFrag/{}'.format(id.hex()), payload)
         assert response.status_code == 200
         return True, ursula.stamp.as_umbral_pubkey()
 
     def send_work_order_payload_to_ursula(self, work_order):
-        mock_client = self.__get_mock_client_by_port(work_order.ursula.rest_interface.port)
+        mock_client = self.__get_mock_client_by_ursula(work_order.ursula)
         payload = work_order.payload()
         id_as_hex = work_order.arrangement_id.hex()
         return mock_client.post('http://localhost/kFrag/{}/reencrypt'.format(id_as_hex), payload)
 
     def get_treasure_map_from_node(self, node, map_id):
-        mock_client = self.__get_mock_client_by_port(node.rest_interface.port)
+        mock_client = self.__get_mock_client_by_ursula(node)
         return mock_client.get("http://localhost/treasure_map/{}".format(map_id))
 
     def node_information(self, host, port):
@@ -229,8 +239,8 @@ class MockRestMiddleware(RestMiddleware):
         response = mock_client.get("http://localhost/public_information")
         return response
 
-    def get_nodes_via_rest(self, address, port, announce_nodes=None, nodes_i_need=None):
-        mock_client = self.__get_mock_client_by_port(port)
+    def get_nodes_via_rest(self, url, announce_nodes=None, nodes_i_need=None):
+        mock_client = self.__get_mock_client_by_url(url)
 
         if nodes_i_need:
             # TODO: This needs to actually do something.
@@ -239,16 +249,16 @@ class MockRestMiddleware(RestMiddleware):
             pass
 
         if announce_nodes:
-            response = mock_client.post("https://{}:{}/node_metadata".format(address, port),
+            response = mock_client.post("https://{}/node_metadata".format(url),
                                      verify=False,
                                      data=bytes().join(bytes(n) for n in announce_nodes))  # TODO: TLS-only.
         else:
-            response = mock_client.get("https://{}:{}/node_metadata".format(address, port),
+            response = mock_client.get("https://{}/node_metadata".format(url),
                                     verify=False)  # TODO: TLS-only.
         return response
 
     def put_treasure_map_on_node(self, node, map_id, map_payload):
-        mock_client = self.__get_mock_client_by_port(node.rest_interface.port)
+        mock_client = self.__get_mock_client_by_ursula(node)
         response = mock_client.post("http://localhost/treasure_map/{}".format(map_id),
                       data=map_payload, verify=False)
         return response
@@ -280,3 +290,35 @@ class MockPolicy(Policy):
                                           expiration=expiration)
 
             self.consider_arrangement(network_middleware=network_middleware, arrangement=arrangement)
+
+
+from twisted.internet import protocol
+
+
+class UrsulaProcessProtocol(protocol.ProcessProtocol):
+
+    def __init__(self, command):
+        self.command = command
+
+    def connectionMade(self):
+        print("connectionMade!")
+        self.transport.closeStdin()   # tell them we're done
+
+    def outReceived(self, data):
+        print(data)
+
+    def errReceived(self, data):
+        print(data)
+
+    def inConnectionLost(self):
+        print("inConnectionLost! stdin is closed! (we probably did it)")
+
+    def outConnectionLost(self):
+        print("outConnectionLost! The child closed their stdout!")
+
+    def errConnectionLost(self):
+        print("errConnectionLost! The child closed their stderr.")
+
+    def processEnded(self, status_object):
+        print("processEnded, status %d" % status_object.value.exitCode)
+        print("quitting")
